Convert items of array paramters
